correcao dano arma

+++ RPG_1/Assets/Scripts/Enemy.cs
@@ -9,7 +9,7 @@ public class Enemy : MonoBehaviour
     public Transform Player;
     public float stoppingDistance;
 
-    public int Health = 50;
+    public float Health = 50;
 
     // Start is called before the first frame update
     void Start()
@@ -27,6 +27,7 @@ public class Enemy : MonoBehaviour
             transform.position = Vector3.MoveTowards(transform.position, Player.position, speed * Time.deltaTime);
         }
 
+       
 
     }
